App.jsx moved to containers
